chore: rename pkg
