+++ cmd/api/api.go
@@ -17,8 +17,8 @@ package main
 import (
 	"log"
 
+	"github.com/bucketeer-io/bucketeer/pkg/api/cmd"
 	"github.com/bucketeer-io/bucketeer/pkg/cli"
-	"github.com/bucketeer-io/bucketeer/pkg/gateway/cmd"
 )
 
 var (
