+++ pkg/api/api/register_event_type.go
