+++ test/e2e/autoops/auto_ops_test.go
@@ -31,12 +31,12 @@ import (
 	"google.golang.org/grpc/status"
 	"google.golang.org/protobuf/encoding/protojson"
 
+	gwapi "github.com/bucketeer-io/bucketeer/pkg/api/api"
+	gatewayclient "github.com/bucketeer-io/bucketeer/pkg/api/client"
 	autoopsclient "github.com/bucketeer-io/bucketeer/pkg/autoops/client"
 	btclient "github.com/bucketeer-io/bucketeer/pkg/batch/client"
 	experimentclient "github.com/bucketeer-io/bucketeer/pkg/experiment/client"
 	featureclient "github.com/bucketeer-io/bucketeer/pkg/feature/client"
-	gwapi "github.com/bucketeer-io/bucketeer/pkg/gateway/api"
-	gatewayclient "github.com/bucketeer-io/bucketeer/pkg/gateway/client"
 	rpcclient "github.com/bucketeer-io/bucketeer/pkg/rpc/client"
 	"github.com/bucketeer-io/bucketeer/pkg/uuid"
 	autoopsproto "github.com/bucketeer-io/bucketeer/proto/autoops"
