+++ test/e2e/eventcounter/eventcounter_test.go
@@ -34,12 +34,12 @@ import (
 	"google.golang.org/grpc/status"
 	"google.golang.org/protobuf/encoding/protojson"
 
+	gwapi "github.com/bucketeer-io/bucketeer/pkg/api/api"
+	gatewayclient "github.com/bucketeer-io/bucketeer/pkg/api/client"
 	btclient "github.com/bucketeer-io/bucketeer/pkg/batch/client"
 	ecclient "github.com/bucketeer-io/bucketeer/pkg/eventcounter/client"
 	experimentclient "github.com/bucketeer-io/bucketeer/pkg/experiment/client"
 	featureclient "github.com/bucketeer-io/bucketeer/pkg/feature/client"
-	gwapi "github.com/bucketeer-io/bucketeer/pkg/gateway/api"
-	gatewayclient "github.com/bucketeer-io/bucketeer/pkg/gateway/client"
 	rpcclient "github.com/bucketeer-io/bucketeer/pkg/rpc/client"
 	"github.com/bucketeer-io/bucketeer/pkg/uuid"
 	btproto "github.com/bucketeer-io/bucketeer/proto/batch"
