+++ test/e2e/util/rest.go
@@ -23,7 +23,7 @@ import (
 	"net/http"
 	"testing"
 
-	gwapi "github.com/bucketeer-io/bucketeer/pkg/gateway/api"
+	gwapi "github.com/bucketeer-io/bucketeer/pkg/api/api"
 	eventproto "github.com/bucketeer-io/bucketeer/proto/event/client"
 	featureproto "github.com/bucketeer-io/bucketeer/proto/feature"
 	gwproto "github.com/bucketeer-io/bucketeer/proto/gateway"
